[Refactor] Move storage workflow tasks to DataWorkflows
