+++ be/src/http/action/checksum_action.cpp
@@ -39,13 +39,13 @@
 
 #include "boost/lexical_cast.hpp"
 #include "common/logging.h"
+#include "data_workflows/consistency/engine_checksum_task.h"
 #include "http/http_channel.h"
 #include "http/http_request.h"
 #include "http/http_status.h"
 #include "runtime/env/global_env.h"
 #include "runtime/exec_env.h"
 #include "runtime/mem_tracker.h"
-#include "storage/task/engine_checksum_task.h"
 
 namespace starrocks {
