+++ be/test/storage/tablet_updates_test.cpp
@@ -23,6 +23,7 @@
 #include "common/config_primary_key_fwd.h"
 #include "common/config_scan_io_fwd.h"
 #include "common/config_storage_fwd.h"
+#include "data_workflows/consistency/engine_checksum_task.h"
 #include "fs/fs_factory.h"
 #include "runtime/runtime_state.h"
 #include "script/script.h"
@@ -31,7 +32,6 @@
 #include "storage/manual_compaction.h"
 #include "storage/primary_key_dump.h"
 #include "storage/rowset/rowset_meta_manager.h"
-#include "storage/task/engine_checksum_task.h"
 #include "storage/txn_manager.h"
 
 namespace starrocks {
